requirements.txt
fastapi
uvicorn
paramiko
python-dotenv
pydantic
python-multipart 